move helpers from pages
